feat(admin-ui):useStringReplace hook and widget titles
